prl-backend/pulse/server/requirements.txt
fastapi
tortoise-orm
asyncmy
mangum
boto3
pydantic
pymysql
httpx
